+++ lib/POST/models/post.dart
@@ -1,0 +1,8 @@
+// 데이터 모델 클래스 정의
+class Post {
+  final String title;    // 글 제목
+  final String content;  // 글 내용
+
+  // 생성자
+  Post(this.title, this.content);
+}
