+++ lib/POST/utils/constants.dart
@@ -1,0 +1,11 @@
+// utils/constants.dart
+
+import 'package:flutter/material.dart';
+
+class AppConstants {
+  // 앱에서 사용될 상수들을 정의하는 클래스입니다.
+
+  // 주요 색상을 정의합니다.
+  static const Color primaryColor = Color.fromRGBO(161, 196, 253, 1);
+  // 다른 상수들을 필요에 따라 추가할 수 있습니다.
+}
